+++ PTD/PTUtils/include/TalkUtil.h
@@ -5,5 +5,5 @@
 
 namespace pt {
     void setMessageArg(TalkMessageCtrl *pTalkCtrl, const wchar_t *pText);
-    void setTextBoxArgStringNumberFontRecursive(LayoutActor* pLayout, const char* paneName, s32 num, s32 index);
+    void setTextBoxArgStringNumberFontRecursive(LayoutActor* pLayout, const char* paneName, u16 num, s32 index);
 };
